% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_NA.R
\name{aprior.na}
\alias{aprior.na}
\alias{bprior.na}
\title{Helper Functions for ComBat.NA}
\usage{
aprior.na(delta.hat)

bprior.na(delta.hat)
}
\arguments{
\item{delta.hat}{multiplicative batch effect}
}
\description{
These functions compute the shape and rate parameters of the inverse gamma
distribution based on the values delta.hat, which are the multiplicative batch effects.

Credit goes to the original authors of the sva package.
}
\section{Functions}{
\itemize{
\item \code{bprior.na}: 
}}

