% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_msnset}
\alias{brain_msnset}
\title{Brain MSnSet}
\format{
An object of class \code{MSnSet} with 2820 rows and 100 columns.
}
\usage{
data("brain_msnset")
}
\description{
An \code{MSnSet} object containing log\eqn{_2} relative
  abundances of peptides detected in brain tissue.
}
\keyword{datasets}
