% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorschemes.R
\name{colorschemes}
\alias{colorschemes}
\alias{blue.colors}
\alias{dusk.colors}
\alias{hot.colors}
\alias{hot2.colors}
\alias{jet.colors}
\alias{jet2.colors}
\alias{turbo.colors}
\title{Favorite Colorschemes}
\usage{
blue.colors(n)

dusk.colors(n)

hot.colors(n)

hot2.colors(n)

jet.colors(n)

jet2.colors(n)

turbo.colors(n, start = 0, end = 1)
}
\arguments{
\item{n}{Number of different colors in the colorscheme}
}
\value{
Vector of colors as hex RGB strings
}
\description{
Provides more colorscheme options in addition to basic ones:
heat.colors, rainbow and so on. Hot and jet colorschemes are
analagous to the MATLAB colorschemes with the same names.
Blue and dusk colorschemes correspond to RColorBrewer "Blues"
and inverse "RdYlBu".
}
\section{Functions}{
\itemize{
\item \code{blue.colors()}: based on RColorBrewer "Blues"

\item \code{dusk.colors()}: based on reverse of RColorBrewer "RdYlBu".
Looks like a dusk.

\item \code{hot.colors()}: based on Matlab's "hot"

\item \code{hot2.colors()}: similar to Matlab's "hot"

\item \code{jet.colors()}: based on Matlab's "jet"

\item \code{jet2.colors()}: similar to Matlab's "jet"

\item \code{turbo.colors()}: Google's(?) turbo color scheme

}}
\examples{

## Linear
image(as.matrix(1:100), col=blue.colors(100))
image(as.matrix(1:100), col=dusk.colors(100))
image(as.matrix(1:100), col=hot.colors(100))
image(as.matrix(1:100), col=hot2.colors(100))
image(as.matrix(1:100), col=jet.colors(100))
image(as.matrix(1:100), col=jet2.colors(100))
image(as.matrix(1:100), col=turbo.colors(100))

## Random
N <- 10
set.seed(0)
mat <- matrix(rnorm(N^2), ncol=N)
image(mat, col=blue.colors(N^2))
image(mat, col=dusk.colors(N^2))
image(mat, col=hot.colors(N^2))
image(mat, col=hot2.colors(N^2))
image(mat, col=jet.colors(N^2))
image(mat, col=jet2.colors(N^2))
image(mat, col=turbo.colors(N^2))

## Airy pattern
x <- seq(-1,+1,length=100)
z <- outer(x,x,function(x,y){sin(10*sqrt(x^2+y^2))})
image(z, col=blue.colors(100))
image(z, col=dusk.colors(100))
image(z, col=hot.colors(100))
image(z, col=hot2.colors(100))
image(z, col=jet.colors(100))
image(z, col=jet2.colors(100))
image(z, col=turbo.colors(100))
}
