% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_heatmap.R
\name{complex_heatmap}
\alias{complex_heatmap}
\title{Annotated Expression or Correlation Heatmaps}
\usage{
complex_heatmap(
  eset,
  clustering_distance = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski", "pearson", "spearman", "kendall"),
  clustering_method = c("ward.D", "ward.D2", "single", "average", "complete", "mcquitty",
    "median", "centroid"),
  heatmap_type = "expression",
  cor_method = c("pearson", "kendall", "spearman"),
  cluster_columns = TRUE,
  cluster_rows = TRUE,
  show_column_dendrogram = TRUE,
  show_row_dendrogram = TRUE,
  show_column_names = FALSE,
  show_row_names = FALSE,
  heatmap_title = character(0),
  heatmap_legend_title = NULL,
  color_range = NULL,
  heatmap_args = list(),
  anno_column = NULL,
  anno_row = NULL,
  anno_column_titles = anno_column,
  anno_row_titles = anno_row,
  anno_column_colors = list(),
  anno_row_colors = list(),
  anno_args = list(),
  filename = NULL,
  height = 5,
  width = 5,
  draw_args = list()
)
}
\arguments{
\item{eset}{an object of class \code{\link[Biobase]{eSet}} or
\code{\link[MSnbase]{MSnSet-class}}.}

\item{clustering_distance}{character; the distance measure used to cluster
rows and columns. Passed to \code{\link[stats]{dist}}. One of
(\code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
\code{"canberra"}, \code{"binary"}, \code{"minkowski"}, \code{"pearson"},
\code{"spearman"}, or \code{"kendall"}). Default is \code{"euclidean"}.}

\item{clustering_method}{character; the agglomeration method used to cluster
rows and columns. Passed to \code{\link[stats]{hclust}}. One of
(\code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"average"},
\code{"complete"}, \code{"mcquitty"}, \code{"median"}, or
\code{"centroid"}). Default is \code{"ward.D"}.}

\item{heatmap_type}{character; the type of heatmap to generate. Must be (an
abbreviation of) either \code{"expression"}, \code{"sample_correlation"},
or \code{"feature_correlation"}. The default (\code{"expression"})
generates a heatmap with features as rows and samples as columns. The other
options calculate the sample or feature correlation matrix and generate a
heatmap with samples or features as both rows and columns.}

\item{cor_method}{character; the method used to generate the correlation
matrix if \code{heatmap_type} is a correlation heatmap. One of
(\code{"pearson"}, \code{"kendall"}, or \code{"spearman"}). Defaults to
\code{"pearson"}. Passed to \code{\link[stats]{cor}}.}

\item{cluster_columns}{logical; whether to cluster the columns.}

\item{cluster_rows}{logical; whether to cluster the rows.}

\item{show_column_dendrogram}{logical; whether to show the column dendrogram.
Does not affect clustering.}

\item{show_row_dendrogram}{similar to \code{show_column_dendrogram}, but for
rows.}

\item{show_column_names}{logical; whether to show the column names.}

\item{show_row_names}{logical; whether to show the row names.}

\item{heatmap_title}{character; overall title for the heatmap.}

\item{heatmap_legend_title}{character; title of the heatmap color legend.}

\item{color_range}{numeric; vector of length 2 used to restrict the colors of
the heatmap body. Useful when color differences are not easily discernible
due to outliers.}

\item{heatmap_args}{list of arguments passed to
\code{\link[ComplexHeatmap]{Heatmap}}.}

\item{anno_column}{character; one or more column names of \code{pData(eset)}
used to annotate the columns of the heatmap. By default, columns are not
annotated.}

\item{anno_row}{character; one or more column names of \code{fData(eset)}
used to annotate the rows of the heatmap. By default, rows are not
annotated.}

\item{anno_column_titles}{character; names for the column annotations, if
different from \code{anno_column}. Must be the same length as
\code{anno_column}.}

\item{anno_row_titles}{similar to \code{anno_column_titles}, but for row
annotations.}

\item{anno_column_colors}{list of custom colors for column annotations
specified by \code{anno_column}. List names must match one or more names in
\code{anno_column}. If modifying the colors of a continuous column
annotation, \strong{always use \code{\link[circlize]{colorRamp2}}} with desired
breaks and colors. Otherwise, pass a vector of unique colors (order
matters).}

\item{anno_row_colors}{same as \code{anno_column_colors}, but for
\code{anno_row}.}

\item{anno_args}{list of arguments passed to
\code{\link[ComplexHeatmap]{HeatmapAnnotation}}.}

\item{filename}{character; file name used to save the heatmap. Must end in
(".png", ".bmp", ".jpg", ".tif", or ".pdf").}

\item{height}{numeric; height of heatmap in inches. Default is \code{5}.}

\item{width}{numeric; width of heatmap in inches. Default is \code{5}.}

\item{draw_args}{list of arguments passed to
\code{\link[ComplexHeatmap]{draw-HeatmapList-method}}. Unlikely to be used
often.}
}
\value{
An object of class \code{\link[ComplexHeatmap]{HeatmapList-class}} or
nothing (save to file instead).
}
\description{
A wrapper around functions from
\code{\link[ComplexHeatmap]{ComplexHeatmap-package}}. Creates expression or
correlation heatmaps from \code{eSet} or \code{MSnSet} objects.
}
\note{
If the \code{hclust} error \verb{NA/NaN/Inf in foreign function call} occurs, it
means there are cases where two or more features are not present in the
same samples, so their distances will be \code{NA}, and it is impossible to
perform hierarchical clustering. If this happens, filter to features
present in more than 50\% of samples with \code{m[rowMeans(!is.na(m)) > 0.5, ]},
where \code{m} is the \code{MSnSet}.
}
\examples{
# library(MSnSet.utils)
data(longitudinal_biomarker_study) # MSnSet
ee <- longitudinal_biomarker_study
# Missingness filter - clustering may fail otherwise
ee <- ee[rowMeans(!is.na(exprs(ee))) >= 0.5, ]

# Expression heatmap
complex_heatmap(ee)

# Limit color range to see differences more easily
complex_heatmap(ee, color_range = c(-2, 2))

# Sample correlation heatmap
complex_heatmap(ee, heatmap_type = "s")

# Annotate columns by "Type" (categorical) and "Age" (continuous)
# Annotate rows by "isSpike" (logical)
complex_heatmap(ee, anno_column = c("Type", "Age"),
                anno_row = "isSpike")

}
\references{
Gu, Z., Eils, R., & Schlesner, M. (2016). Complex heatmaps reveal
patterns and correlations in multidimensional genomic data.
\emph{Bioinformatics}, \emph{32}(18), 2847-2849.
\url{https://doi.org/10.1093/bioinformatics/btw313}

Gu, Z., Gu, L., Eils, R., Schlesner, M., & Brors, B. (2014). Circlize
implements and enhances circular visualization in R. \emph{Bioinformatics},
\emph{30}(19), 2811-2812. \url{https://doi.org/10.1093/bioinformatics/btu393}
}
\seealso{
\href{https://jokergoo.github.io/ComplexHeatmap-reference/book/}{ComplexHeatmap
Complete Reference}
}
