% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_by_features.R
\name{concatenate_by_features}
\alias{concatenate_by_features}
\title{Concatenate MSnSet Objects by Features}
\usage{
concatenate_by_features(..., tags = NULL)
}
\arguments{
\item{...}{MSnSet objects}

\item{tags}{vector of characters attached to feature names of
corresponding MSnSet objects}
}
\value{
MSnSet object with concatenated features
}
\description{
Combines MSnset objects by concatenating features. Samples must
be the same and maintained the same. Pheno data also must be the
same.
}
\examples{
data(srm_msnset)
dim(msnset)
featureNames(msnset)
msnset3 <- msnset2 <- msnset
# Z-score transform
exprs(msnset2) <- sweep(exprs(msnset2), 1, apply(exprs(msnset2), 1, sd), '/')
# retaining just the sign
exprs(msnset3) <- ifelse(exprs(msnset3) > 0, 1, -1)
m <- concatenate_by_features(msnset, msnset2, msnset3,
                             tags=c("_ori","_zsc","_sgn"))
dim(m)
featureNames(m)

}
