% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_dist.R
\name{dist}
\alias{dist}
\title{Customized Distance Function}
\usage{
dist(x, method = "euclidean", na.avg = FALSE, ...)
}
\arguments{
\item{x}{a numeric matrix}

\item{method}{the distance measure to be used. 
This must be one of "euclidean", "maximum", "manhattan", 
                    "canberra", "binary", "minkowski",
                    "pearson", "kendall", "spearman".
Any unambiguous substring can be given.}

\item{na.avg}{logical if yes, then NA distance will 
be imputed with average distance
value. Default is FALSE.}

\item{...}{arguments of \code{\link[stats]{dist}}}
}
\description{
Based on \code{\link[stats]{dist}}. Additionally
it can handle correlation-based distances and 
has argument for imputation of NA distances. To
avoid confusion with \code{stats::dist} all the
default settings are such that it performs equivalently
to the aforementioned one.
}
\note{
In the current set-up
     corrlation values -1, 0 and 1 correspond to
     distances 1, 0.5 and 1, respectively.\cr
     R=1  -> dist=0 \cr
     R=0  -> dist=0.5 \cr
     R=-1 -> dist=1 \cr
}
\examples{
x <- matrix(c(1:5, 5:9, 9:13), ncol=3)
stats::dist(x)
dist(x)
dist(x, method="pearson")

}

