% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eset_lm_1.R
\name{eset_lm}
\alias{eset_lm}
\title{Nested Linear Model Analysis}
\usage{
eset_lm(eset, form.alt, form.nul, facs = NULL, norm.coef = NULL)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{form.alt}{character formulation of alternative model}

\item{form.nul}{character formulation of NULL model}

\item{facs}{data frame with the factors in its columns.
If NULL, then pData(eset) will be used.}

\item{norm.coef}{vector with sample-to-sample normalization
coefficients in log2 scale.}
}
\value{
data.frame
     \describe{
         \item{\code{effect}}{either max contrast in case of factor
                              or slope in case of continuous variable}
         \item{\code{F.stat}}{F statistic}
         \item{\code{p.value}}{p-value}
     }
}
\description{
A function for analysis of sigfinicance of factors (akin ANOVA)
or continuous variable (regression) using nested linear models.
The function modelled after `msmsTests::msms.glm.qlll`
}

