% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/presense_absense.R
\name{eset_presence_absence}
\alias{eset_presence_absence}
\title{Presence/Absence Testing}
\usage{
eset_presence_absence(eset, grouping, test = c("fisher", "chisq", "g",
  "binom"), ...)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{grouping}{character defining the column in phenoData}

\item{test}{character: Fisher exact test, \eqn{\chi^{2}}{\chi^2}, G-test or binomial.}

\item{...}{other aruments to pass to the test functions
(e.g. simulate.p.value = TRUE)}
}
\value{
data.frame
     \describe{
         \item{\code{p.value}}{p-value}
         \item{\code{effect}}{difference between second and first group
                             (according to the factor levels).
                             In case of larger number of levels,
                             it is the absolute value of the max diffence}
         \item{\code{...}}{proportions of present for each group.
                             Column names are level names.}
     }
}
\description{
A function for testing the significance in differences
of presence/absence patterns.
}
\details{
The recommended way of using the test is two-group comparison.
All except "binom" options will work for larger number of groups.
}
\examples{
library("MSnbase")
Nsam = 20
Npep = 5
M <- matrix(rbinom(Nsam*Npep, 1, 0.5), ncol=Nsam)
pd <- data.frame(group = gl(2, Nsam/2))
fd <- data.frame(otherfdata = letters[1:Npep])
x0 <- MSnSet(M, fd, pd)
eset_presence_absence(x0, 'group', test='fisher')
eset_presence_absence(x0, 'group', test='chisq', simulate.p.value=TRUE)
eset_presence_absence(x0, 'group', test='g')
eset_presence_absence(x0, 'group', test='binom')
}

