% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustive_ratios.R
\name{exhaustive_ratios}
\alias{exhaustive_ratios}
\title{Exhaustive Ratios}
\usage{
exhaustive_ratios(m, INDEX = NULL)
}
\arguments{
\item{m}{MSnSet object}

\item{INDEX}{character Name of one of the columns in fData(m). 
If NULL (defaults), then ratios generated assuming all
features (e.g. peptides or proteforms) belong to one
grouping factor.}
}
\description{
For a group of features size N it exhaustively
generates N*(N-1)/2 ratios and returns another MSnSet.
}
\details{
There are few scenarios when this function can be useful.
\itemize{
     \item SRM data. Generate full exhaustive ratios to normalized per
         tissue leakage.
     \item Top-Down data. Generate ratios of proteoforms within a given gene.
     \item any data. Ratios within a given pathway to assess dysregulation.
}
}

