% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_occurrence.R
\name{filter_by_occurrence}
\alias{filter_by_occurrence}
\title{Retain features passing occurrence threshold}
\usage{
filter_by_occurrence(m, occurrence = 0.5)
}
\arguments{
\item{m}{ExpresionSet or MSnSet object}

\item{occurrence}{Minimum proportion of observations}
}
\value{
(MSnSet) MSnSet object
}
\description{
Retain features passing occurrence threshold
}
\examples{
#
data(srm_msnset)
MSnSet.utils::image_msnset(msnset)
msnset2 <- filter_by_occurrence(msnset, occurrence = 0.6)
MSnSet.utils::image_msnset(msnset2)
}
