% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.3.R
\name{heatmap.3}
\alias{heatmap.3}
\title{Heatmap}
\usage{
heatmap.3(x, dist.fun = c("euclidean", "pearson"), na.avg = TRUE,
  col.ramp = bluered, z.transf = c(FALSE, TRUE), breaks = seq(-1, +1,
  length = 100), linkage = c("average", "ward.D", "ward.D2", "single",
  "complete", "mcquitty", "median", "centroid"), nLabLim = 30, ...)
}
\arguments{
\item{x}{MSnSet object}

\item{dist.fun}{distance function "euclidean" or"pearson"}

\item{na.avg}{logical. Should NA distances be assigned just an average value?}

\item{col.ramp}{color mapping function. default gplots::bluered}

\item{z.transf}{logical perform Z-transform or not.}

\item{breaks}{color key breaks}

\item{linkage}{see ?hclust}

\item{nLabLim}{max limit of the row/column labels to show. default 30}

\item{...}{further arguments to \code{gplots::heatmap.2}}
}
\description{
Customized Heatmap. TO FILL WHAT EXACTLY IS CUSTOM ABOUT IT.
}
\examples{
data(srm_msnset)
set.seed(0)
clrz <- sample(colors(), 17)
heatmap.3(cor(exprs(msnset)), 
          dist.fun="pearson",
          linkage="average", 
          nLabLim=50,
          ColSideColors=c('red3','yellow3','green3')[as.factor(pData(msnset)$subject.type)],
          RowSideColors=clrz[as.factor(pData(msnset)$match.group)])
}

