% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.plus.2.R
\name{heatmap.plus.2}
\alias{heatmap.plus.2}
\title{Tweak of "heatmap.plus"}
\usage{
heatmap.plus.2(x, Rowv = NULL, Colv = if (symm) "Rowv" else NULL,
  distfun = dist, hclustfun = hclust, reorderfun = function(d, w)
  reorder(d, w), add.expr, symm = FALSE, revC = identical(Colv, "Rowv"),
  scale = c("row", "column", "none"), na.rm = TRUE, margins = c(5, 5),
  ColSideColors, RowSideColors, propColSide = 0.1, propRowSide = 0.1,
  cexRow = 0.2 + 1/log10(nr), cexCol = 0.2 + 1/log10(nc), labRow = NULL,
  labCol = NULL, main = NULL, xlab = NULL, ylab = NULL,
  keep.dendro = FALSE, verbose = getOption("verbose"), ...)
}
\arguments{
\item{propColSide}{numeric controls how much room to allocate for the
side color bars. E.g. 0.5 is even split between the
heatmap and the bars. Default is 0.1. That is 10\%
bars, 90\% heatmap.}

\item{propRowSide}{numeric see above}
}
\description{
This function is basically \code{\link[heatmap.plus]{heatmap.plus}} 
from \pkg{heatmap.plus} package.
Except with a pair of additional arguments controling the proportion
of row and column color sidebars.
}
\examples{
z = matrix(rnorm(30),nrow=5,ncol=6);
rlab = matrix(as.character(c(1:5,2:6,3:7,4:8)),nrow=5,ncol=4);
clab = matrix(as.character(c(1:6,6:1)),nrow=6,ncol=2);
colnames(rlab) = LETTERS[1:dim(rlab)[2]];
colnames(clab) = 1:dim(clab)[2];
heatmap.plus.2(z, ColSideColors=clab, 
                  RowSideColors=rlab,
                  propColSide=0.1,
                  propRowSide=0.2);
}

