% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ids}
\alias{ids}
\title{Rat Phospho Identifications}
\format{
An object of class \code{data.frame} with 19938 rows and 3 columns.
}
\usage{
data("phospho_identifications_rat")
}
\description{
Description goes here.
}
\keyword{datasets}
