% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.r
\name{knn_imputation}
\alias{knn_imputation}
\title{Customized K-NN Imputation}
\usage{
knn_imputation(x, K = 10, show.diagnostics = F)
}
\arguments{
\item{x}{MSnSet or ExpressionSet object}

\item{K}{number of nearest neighbors}

\item{show.diagnostics}{logical indicating if to 
plot the results of imputation for each feature}
}
\description{
K-NN imputation adopted to LC-MS proteomics data.
The main reason for missing data in LC-MS datasets is
low abundance of the protein/peptides. Therefore this 
K-NN imputation algorithm explicitely relies on this
assumption.
}
\details{
The algorithm. For each row in the exprs matrix, 
that contain missing values, perform the following steps:
\enumerate{
 \item impute missing values with the lowest values in the row (feature)
 \item find K features (with no missing values) with 
         highest Spearman correlation
 \item scale the K-neighbors, so that median intensity ratio is 1
 \item impute missing values with mean value of scaled K-neighbors
}
}
\note{
The algorithm assumes that the data is not log-transformed. Thus, if
         the data is log-transform - exponentiate.
}
\examples{
suppressPackageStartupMessages(library("MSnbase"))
data(naset)
image(naset[1:50,])
x <- knn_imputation(naset)
image(x[1:50,])

}
