% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{layout_compact_stack}
\alias{layout_compact_stack}
\title{Compact stacking of peptides}
\usage{
layout_compact_stack(prot, min_y = 0, width_y = 0.025, step_y = 0.033)
}
\arguments{
\item{prot}{`data.frame` containing peptides from a single protein with
columns "First_AA" and "Last_AA".}

\item{min_y}{numeric; shifts peptides vertically.}

\item{width_y}{numeric; vertical width of each peptide. Must be less than
`step_y`.}

\item{step_y}{numeric; step size to resolve peptide overlaps.}
}
\description{
Compact stacking of peptides
}
