% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{layout_manhattan}
\alias{layout_manhattan}
\title{Manhattan-style arrangement of peptides}
\usage{
layout_manhattan(prot, p_val_from, min_y = 0, width_y = 0.025)
}
\arguments{
\item{prot}{`data.frame` containing peptides from a single protein.}

\item{p_val_from}{character; name of a column in `prot` providing statistical
significance of each peptide.}

\item{min_y}{numeric; shifts peptides vertically.}

\item{width_y}{numeric; vertical width of each peptide rectangle.}
}
\description{
Manhattan-style arrangement of peptides
}
\details{
Plots peptides according to their statistical significance.
  Specifically, this function will vertically-center peptides according to
  their -log10(`p_val_from`) scaled to the largest value or a value of 5.
}
