% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limma_a_b.R
\name{limma_a_b}
\alias{limma_a_b}
\title{Wrapper for Limma}
\usage{
limma_a_b(eset, model.str, coef.str)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{model.str}{character formulation of the model (e.g. "~ a + b")}

\item{coef.str}{character coefficient of interest. E.g. "a".
Either numeric or factor with 2 levels.}
}
\value{
data.frame. Basically output of
     \code{\link[limma]{topTable}} function.
}
\description{
A convenience wrapper for limma
}
\examples{
library("vp.misc")
data("cptac_oca")
library("limma")
ee <- oca.set
ee <- ee[rowSums(!is.na(exprs(ee))) >= 30,]
model.string <- "~ PLATINUM.STATUS + AGE" # model with covariates
coef.string <- "PLATINUM.STATUS" # note, only two levels allowed here

res <- limma_a_b(oca.set,
                 model.str = "~ PLATINUM.STATUS + AGE",
                 coef.str = "PLATINUM.STATUS")
head(res)
}

