% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_contrasts.R
\name{limma_contrasts}
\alias{limma_contrasts}
\title{Wrapper for Limma}
\usage{
limma_contrasts(
  eset,
  coef.str,
  covariates = NULL,
  contrasts,
  contrast.sep = "_",
  adjust.method = "BH",
  ...
)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{coef.str}{vector of character coefficients of interest from
\code{colnames(pData(eset))}. E.g. "a" or c("a", "b").}

\item{covariates}{character vector of potential covariates that
should be included in the model.}

\item{contrasts}{character vector of contrasts to test.}

\item{contrast.sep}{string used to join the level labels of the columns
specified by \code{coef.str}. The default is an underscore, so factor A
with entries \code{c("a1", "a1", "a2")} and factor B with entries
\code{c("b1", "b2", "b3")} would combine to form a factor with entries
\code{c("a1_b1", "a1_b2", "a2_b3")}.}

\item{adjust.method}{method for p-value adjustment. Default is "BH"
(Benjamini-Hochberg).}

\item{...}{arguments for \code{\link[limma]{lmFit}}}

\item{model.str}{character formulation of the model (e.g. "~ 0 + a + b")}

\item{ref.str}{(character) Reference string for pairwise comparisons}
}
\value{
data.frame. Basically output of \code{\link[limma]{topTable}}
     function with additional columns \code{feature} and \code{contrast}.
}
\description{
A convenience wrapper for limma contrast testing.
}
\examples{

library(MSnSet.utils)
data(cptac_oca)

coef.str <- c("SUBTYPE", "SURVIVALSTATUS")
covariates <- c("AGE", "Batch")

# Testing a single contrast
contrasts <- c("Immunoreactive_LIVING - Immunoreactive_DECEASED")

res <- limma_contrasts(oca.set,
                       coef.str = coef.str,
                       covariates = covariates,
                       contrasts = contrasts)

head(res)
}
