% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2_zero_center.R
\name{log2_zero_center}
\alias{log2_zero_center}
\title{Log2 transform and center around zero}
\usage{
log2_zero_center(m)
}
\arguments{
\item{m}{ExpresionSet or MSnSet object}
}
\value{
(MSnSet) MSnSet object
}
\description{
Log2 transform applied to expression crosstab of ExpressionSet or MSnSet
object followed by centering the median around zero. Converts one
MSnSet to another MSnSet.
}
\examples{
data(srm_msnset)
msnset2 <- log2_zero_center(msnset)
}
