% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal_biomarker_study}
\alias{longitudinal_biomarker_study}
\title{Example of Longitudinal Biomarker Study}
\format{
a \code{MSnSet} instance
}
\source{
xxx study
}
\description{
Targeted SRM analysis. Time points A,B,C and D 
increase in the corresponding order. A & B are prior the disease diagnosis 
for the cases. C & D are correspondingly after the diagnosis. Time point D
is after the treatment.
}
\author{
Vlad Petyuk, 2018-11-09
}
