% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_objects.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{arguments passed to \code{.ls.objects}.}

\item{n}{number of rows to output}
}
\description{
Description goes here.
}
