% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.unique.2.R
\name{make.unique.2}
\alias{make.unique.2}
\title{Custom Make Unique}
\usage{
make.unique.2(x)
}
\arguments{
\item{x}{character vector with non-unique entries}
}
\value{
character vector where all entries have attached suffix
         resolving their non-uniqueness.
}
\description{
Similar to \code{\link[base]{make.unique}}, but with some
customization. E.g. starts counting with ".1" instead of blank.
}
\examples{
make.unique(c('a','b','b'))
make.unique.2(c('a','b','b'))
}
