% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects_anova.R
\name{mixed_effects_anova}
\alias{mixed_effects_anova}
\title{Mixed Effects ANOVA}
\usage{
mixed_effects_anova(m, alt_model, test_factor, order_by = c("none", "p-value"))
}
\arguments{
\item{m}{ExpressionSet or MSnSet object}

\item{alt_model}{character formulation of alternative model}

\item{test_factor}{name of the tested factor}

\item{order_by}{"none" or "p-value". Default is "none".}
}
\value{
data.frame
     \describe{
         \item{\code{feature}}{protein or feature names in general.
         Should be as in sampleNames(m).}
         \item{\code{p.val}}{ANOVA p-value}
         \item{\code{p.adj}}{Benjamini-Hochberg adjusted p-value}
         \item{\code{q.val}}{q-value.
         Parameters for qvalue() are: lambda=0.05, pi0.method="bootstrap"}
     }
}
\description{
A wrapper around \code{\link[lme4]{lmer}} functionality.
}
\examples{
data(srm_msnset)
# -- make sure the factors are factors (not characters or integers)
# This is intentionally left for user's control.
msnset$subject.type <- as.factor(msnset$subject.type)
msnset$match.group <- as.factor(msnset$match.group)
msnset$PrepBatch <- as.factor(msnset$PrepBatch)
# -- model description
# subject.type - fixed effect
# match.group - random effect affecting the intercept
# PlateCol - another random effect influencing intercept
out <- mixed_effects_anova(msnset,
                           alt_model="~ subject.type + (1|match.group) + (1|PrepBatch)",
                           test_factor="subject.type",
                           order_by="p-value")
head(out)
}
