% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msnset}
\alias{msnset}
\title{SRM proteomics data from human CSF}
\format{a \code{MSnSet} instance}
\description{
Targeted SRM analysis of 33 proteins in CSF
}
\author{
Vlad Petyuk, 2014-03-24
}

