% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnset2txt.R
\name{msnset2txt}
\alias{msnset2txt}
\title{Exporting MSnSet as text file}
\usage{
msnset2txt(m, prefix = NULL, sig.dig = 3, drop = NULL)
}
\arguments{
\item{m}{MSnSet object}

\item{prefix}{character. All three text file names will 
start with this prefix. By default, the R object name.}

\item{sig.dig}{number of significant digits in expression data.
Default is 3.}
}
\description{
Writes MSnSet as three tables: expression, features and phenotypes.
The format is tab-delimited text.
}
\examples{
\dontrun{
data(srm_msnset)
msnset2txt(msnset)
}
}

