% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/norm_by_global.R
\name{normalizeByGlob}
\alias{normalizeByGlob}
\title{Normalization of SRM Data.}
\usage{
normalizeByGlob(m)
}
\arguments{
\item{m}{MSnSet object}
}
\description{
Normalizes data based on assumption that on average peptide deviation
is zero.  The approach is typically used for global shotgun proteomics
data. Converts one MSnSet to another MSnSet. Uses median polish algorithm.
}

