% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_by_ref.R
\name{normalizeByRef}
\alias{normalizeByRef}
\title{Normalization of SRM Data}
\usage{
normalizeByRef(m, exprsn)
}
\arguments{
\item{m}{MSnSet object}

\item{exprsn}{expression with proteins as terms.
It is evaluated and used as reference.}
}
\description{
Subtracts expression data of one condition from another. 
Converts one MSnSet to another MSnSet.
}
