% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{oca.set}
\alias{oca.set}
\title{CPTAC Ovarian Cancer Proteomics Dataset}
\format{an \code{MSnID} instance.  8103 proteins by 73 features.}
\source{
CPTAC study
}
\description{
Processed iTRAQ data from both PNNL and JHU sites.
}
\examples{
data(cptac_oca)
head(pData(oca.set))
head(fData(oca.set))
head(exprs(oca.set))
}

