% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paircomp.R
\name{paircomp}
\alias{paircomp}
\title{Construct Pairwise Comparisons}
\usage{
paircomp(x, name = "", ref = "", sep = "-")
}
\arguments{
\item{x}{character vector or factor.}

\item{name}{character appended before each element of x when constructing
the contrasts. Default is \code{""}.}

\item{ref}{character; value of x that will be used to compare all other
values against. Creates reference comparisons instead of all pairwise
comparisons.}

\item{sep}{character used to separate terms in the comparisons. The default,
\code{"-"}, generates comparisons of the form \code{"a-b"}.}
}
\value{
character vector of comparisons
}
\description{
Construct a vector of comparisons of the form \code{namea-nameb},
where \code{a} and \code{b} are unique elements of the factor or character
vector \code{x}, and \code{name} is a character string.
}
\details{
\code{paircomp} is primarily intended to generate contrasts for
\code{\link[MSnSet.utils]{limma_contrasts}}. If using \code{paircomp} for
this purpose, \code{x} will be a column in \code{pData(eset)} and \code{name}
will be the name of that column.
}
\examples{
library(MSnSet.utils)

# Character vector for comparisons
x <- LETTERS[1:5]

# All pairwise comparisons
paircomp(x)

# Compare against specific level
paircomp(x, ref = "B")

# Compare against specific level and append name
paircomp(x, name = "test", ref = "B")
}
