% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_contrasts.R
\name{pairwise_contrasts}
\alias{pairwise_contrasts}
\title{Construct Vector of Contrasts}
\usage{
pairwise_contrasts(x, fct, ref.str = NULL, make_names = TRUE, sep = "-")
}
\arguments{
\item{x}{matrix or data frame that contains the column specified
by \code{fct}.}

\item{fct}{character or factor variable from which contrasts will
be generated. If \code{fct} is not a factor, it will be converted
to one.}

\item{ref.str}{level(s) of \code{fct} that all other levels should
be compared against. This is a character vector.}

\item{make_names}{if \code{TRUE} (the default), the name of \code{fct}
will be appended to the beginning of each level of \code{fct} in
the contrasts. This is required for
\code{\link[MSnSet.utils]{limma_contrasts}}.}

\item{sep}{string used to separate terms in the contrasts. The default,
\code{"-"}, generates contrasts of the form \code{"a-b"}.}
}
\value{
character vector of contrasts.
}
\description{
Construct a vector of contrasts of the form \code{a-b}, where \code{a} and
\code{b} are factor levels.
}
\examples{
# Load package
library(MSnSet.utils)

# Data frame for testing
x <- data.frame(v1 = c("a", "b", NA, "c", "d"),
                v2 = c("e", "e", "f", "f", "g"))
x$v3 <- ifelse(is.na(x$v1) | is.na(x$v2), NA,
               paste(x$v1, x$v2, sep = "_"))

pairwise_contrasts(x, "v1", make_names = FALSE)
pairwise_contrasts(x, "v2")
pairwise_contrasts(x, "v2", ref.str = "g",
                   make_names = FALSE)
pairwise_contrasts(x, "v3", ref.str = c("b_e", "a_e"),
                   make_names = FALSE)
}
