% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_reduction.R
\name{pc_reduction}
\alias{pc_reduction}
\title{Principal Component Reduction}
\usage{
pc_reduction(eset)
}
\arguments{
\item{eset}{Expression-Set object or similar}
}
\description{
Reduce an ExpressionSet object down to its principal components.
}
\examples{
data("cptac_oca")
oca.pc <- pc_reduction(oca.set)

res <- limma_a_b(oca.pc, 
                 model.str = "~ PLATINUM.STATUS + AGE", 
                 coef.str = "PLATINUM.STATUS")

}
