% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA.R
\name{PeCorA}
\alias{PeCorA}
\title{Peptide Correlation Analysis (PeCorA)}
\usage{
PeCorA(
  m,
  proteins,
  treatment,
  control_group = NULL,
  protein_column = "Protein",
  peptide_column = "Peptide",
  standardize_samples = TRUE,
  center_peptides = TRUE,
  median_mod = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{m}{\code{MSnSet} object with pData and fData tables. fData must have
columns \code{protein_column} and \code{peptide_column}. The exprs data
should be aggregated such that features are unique combinations of peptide
and protein.}

\item{proteins}{character; one or more proteins in
\code{fData(m)[[protein_column]]} with peptides to test. Only those
proteins with at least 2 peptides will be considered.}

\item{treatment}{character; the name of column in \code{pData(m)} containing
treatment information. The data can be numeric or categorical.
\code{PeCorA} will test for an interaction between \code{treatment} and all
peptides that map to the proteins in \code{proteins}.}

\item{control_group}{character; optional control group. Only applicable if
\code{treatment} is categorical. See \code{center_peptides} parameter below
for details.}

\item{protein_column}{character; the name of a column in \code{fData(m)}
containing protein information. Default is "Protein".}

\item{peptide_column}{character; the name of a column in \code{fData(m)}
containing peptide information. Default is "Peptide".}

\item{standardize_samples}{logical; whether to standardize samples (columns)
in exprs data.}

\item{center_peptides}{logical; whether to mean-center each peptide. If the
data in the \code{treatment} column is categorical and \code{control_group}
is not \code{NULL}, peptides will be zero-centered on the control group
mean. Otherwise, the mean of all samples are used for centering.}

\item{median_mod}{logical; whether to use a modified version of PeCorA. See
details.}

\item{verbose}{logical; whether to produce progress messages.}
}
\value{
An \code{MSnSet} where the fData table includes new or updated
  columns "pecora_pval_{treatment}" and "pecora_adj_pval_{treatment}".
}
\description{
Perform Peptide Correlation Analysis (PeCorA) as described in
  \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/}.
}
\details{
By default, the MSnSet provided by \code{m} will be normalized as
  described in the PeCorA publication: standardize columns and mean-center
  peptides using the mean of the \code{control_group} samples (if provided)
  or the mean of all samples.

  If \code{median_mod = TRUE}, the intensities of "all other peptides" will
  be aggregated to their per-sample medians. This way, all peptides mapping
  to a given protein (aside from the peptide of interest as the function
  loops over all peptides) are not treated as 'independent' observations. The
  effect is that p-values are much higher (less significant) than in the
  original version of PeCorA.

  This function (as well as any helper functions) was based on those from the
  PeCorA R package (\url{https://github.com/jessegmeyerlab/PeCorA}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
\seealso{
\code{\link[MSnSet.utils]{plot_PeCorA}}
}
