% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_analysis.R
\name{pecora_analysis}
\alias{pecora_analysis}
\title{Peptide Correlation Analysis (PeCorA)}
\usage{
pecora_analysis(m, treatment, proteins, median_mod = FALSE)
}
\arguments{
\item{m}{\code{MSnSet} object with pData and fData tables. fData must have
columns "Protein" and "Peptide".}

\item{treatment}{character; the name of the pData column containing treatment
information. The data can be numeric or categorical.}

\item{proteins}{character; one or more proteins in
\code{fData(m)[["Protein"]]} to analyze.}

\item{median_mod}{logical; whether to use the median version of PeCorA. See
\code{\link[MSnSet.utils]{pecora_mod}} for details.}
}
\value{
An \code{MSnSet} where the fData table includes new or updated
  columns "pecora_pval_{treatment}" and "pecora_adj_pval_{treatment}".
}
\description{
Perform Peptide Correlation Analysis (PeCorA) as described in
  \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/}.
}
\details{
The MSnSet provided to \code{m} should be normalized as described in
  the PeCorA publication. See \code{\link[MSnSet.utils]{pecora_preprocess}}
  for details. This function was based on the \code{PeCorA_analysis} function
  from the PeCorA R package
  (\url{https://github.com/jessegmeyerlab/PeCorA}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
\seealso{
\code{\link[MSnSet.utils]{pecora_preprocess}},
  \code{\link[MSnSet.utils]{pecora_mod}},
  \code{\link[MSnSet.utils]{plot_pecora}}
}
