% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_mod.R
\name{pecora_mod}
\alias{pecora_mod}
\title{Modified Peptide Correlation Analysis (PeCorA)}
\usage{
pecora_mod(x, median_mod = FALSE)
}
\arguments{
\item{x}{PeCorA input table. See the original PeCorA function for details.}

\item{median_mod}{logical; whether to use the median version of PeCorA
(\code{median_mod = TRUE}). Default is \code{FALSE}, which will use the
original version.}
}
\value{
A \code{data.frame} with columns "Protein", "Peptide", "pvalue", and
  "adj_pval".
}
\description{
This is a modified version of the PeCorA function. For the
  original paper see \url{https://doi.org/10.1021/acs.jproteome.0c00602}.

  The modification involves aggregating the intensities of "all other
  peptides" within the test described in the PeCorA paper. We aggregate "all
  other peptides" by sample using the median. This way, each peptide mapping
  to a particular protein is not treated as an 'independent' observation. The
  effect is that p-values are much higher (less significant) than in the
  original version of PeCorA.

  This median modification is turned off by default, and only affects the
  p-values and adjusted p-values.
}
\details{
p-values are adjusted across all peptides for a given protein using
  the method of Benjamini and Hochberg.

  This function was based on the \code{PeCorA} function from the PeCorA R
  package (\url{https://github.com/jessegmeyerlab/PeCorA}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
