% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pecora_msnset}
\alias{pecora_msnset}
\title{PeCorA MSnSet}
\format{
An object of class \code{MSnSet} with 2820 rows and 100 columns.
}
\usage{
data("pecora_msnset")
}
\description{
An \code{MSnSet} object to test the PeCorA functions.
}
\keyword{datasets}
