% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_analysis.R
\name{pecora_plot}
\alias{pecora_plot}
\title{Plots a particular peptide using the PeCorA input and output tables.}
\usage{
pecora_plot(m, chosen_feature, treatment_string = NULL, median_mod = FALSE)
}
\arguments{
\item{m}{msnset object containing a pData AND fData table. fData must have a Protein and Peptide column.}

\item{pecora_results}{PeCorA results table from pecora_analysis.}

\item{chosen_protein}{The protein of interest.}

\item{chosen_peptide}{The peptide of interest. Should map to the supplied protein.}
}
\value{
Plot of the specified peptide + protein.
}
\description{
Plots a particular peptide using the PeCorA input and output tables.
}
