% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_analysis.R
\name{pecora_preprocess}
\alias{pecora_preprocess}
\title{This is more of a convenience function, used to normalize as was done in the original PeCorA paper.
Pass an msnset here to preprocess the data. The result is again an msnset. Can perform two normalization
steps, also carried out in the PeCorA paper. Use the parameters peptide_standardize and sample_standardize
to select which steps should be performed.}
\usage{
pecora_preprocess(
  m,
  treatment_string,
  control_group = NULL,
  sample_standardize = FALSE,
  peptide_standardize = FALSE
)
}
\arguments{
\item{m}{msnset object containing a pData AND fData table. fData must have a Protein and Peptide column.}

\item{treatment_string}{the name of the column in pData containing treatment information. The data can be numerical of categorical.}

\item{sample_standardize}{logical for whether exprs data should be standardized so that sample wise we have mean zero and variance one.}

\item{peptide_standardize}{logical for whether exprs data should be normalized so that logratio data is relative to the control group. When using numerical data, this determines whether to zero center each peptide.}
}
\value{
PeCorA results table for the supplied proteins. Saves plots of the significant peptides to the given folder.
}
\description{
If it's desired that both normalization steps be skipped (both set to FALSE), then pecora_preprocess does nothing
at all to the data, and this step can be safely skipped.
}
