% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptide_count_data}
\alias{peptide_count_data}
\title{Peptide Count MSnSet}
\format{
An object of class \code{MSnSet} with 4923 rows and 18 columns.
}
\usage{
data("peptide_count_data")
}
\description{
An \code{MSnSet} object with 4923 features and 18 samples.
}
\keyword{datasets}
