% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{peptide_significance_map_continuous}
\alias{peptide_significance_map_continuous}
\title{Plot peptides on protein sequence with gradient fill}
\usage{
peptide_significance_map_continuous(
  x,
  p_val_from,
  min_y = 0,
  width_y = 0.025,
  step_y = 0.033,
  border_color = "white",
  border_size = NULL,
  aa_step = 20
)
}
\arguments{
\item{x}{`data.frame`.}

\item{p_val_from}{character; name of a column in `x` containing p-values.}

\item{min_y}{numeric; shifts peptides vertically.}

\item{width_y}{numeric; vertical width of each peptide rectangle.}

\item{step_y}{numeric; step size to resolve peptide overlaps.}

\item{border_color}{character; color used for the borders of the peptide
rectangles.}

\item{border_size}{numeric or `NULL`; linewidth of the borders around the
peptide rectangles. If `NULL` (default), the linewidth will be calculated
from the protein length.}

\item{aa_step}{numeric; distance between x axis breaks, starting at `aa_step`
and continuing along the length of the protein.}
}
\description{
Plot peptides on protein sequence with gradient fill
}
\details{
A gradient fill dependent on the values of `p_val_from` is used for
  the peptide rectangles.
}
