% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{peptide_significance_map_manhattan}
\alias{peptide_significance_map_manhattan}
\title{Manhattan arrangement of peptides with p-value fill}
\usage{
peptide_significance_map_manhattan(
  prot,
  p_val_from,
  fill_by = NULL,
  min_y = 0,
  width_y = 0.025,
  border_color = "white",
  border_size = NULL,
  aa_step = 20
)
}
\arguments{
\item{prot}{`data.frame` with a column for p-values and all columns required
by `layout_manhattan`.}

\item{p_val_from}{character; name of a column in `prot` containing p-values.}

\item{fill_by}{character; name of a column in `prot` used to fill the peptide
rectangles. Default is `NULL`, which will use `p_val_from`.}

\item{min_y}{numeric; shifts peptides vertically.}

\item{width_y}{numeric; vertical width of each peptide rectangle.}

\item{border_color}{character; color used for the borders of the peptide
rectangles.}

\item{border_size}{numeric or `NULL`; linewidth of the borders around the
peptide rectangles. If `NULL` (default), the linewidth will be calculated
from the protein length.}

\item{aa_step}{numeric; distance between x axis breaks, starting at `aa_step`
and continuing along the length of the protein.}
}
\description{
Manhattan arrangement of peptides with p-value fill
}
