% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annHeatmap2.R
\name{plot_annHeatmap2}
\alias{plot_annHeatmap2}
\title{Tweaks of plot.annHeatmap}
\usage{
plot_annHeatmap2(x, na.color = "lightgrey", widths, heights, ...)
}
\arguments{
\item{x}{an object of class \code{annHeatmap}.}

\item{na.color}{character color of NA values}

\item{widths}{a numerical vector giving the widths of the sub-plots currently
defined.}

\item{heights}{a numerical vector giving the heights of the sub-plots
currently defined}

\item{...}{additional arguments passed to \code{\link[graphics]{image}}.}
}
\description{
Tweaks of \code{\link[Heatplus]{plot.annHeatmap}} method for
\code{annHeatmap} class.  Additional argument for defining color
for missing values.  There is also a hack of \code{\link[Heatplus]{picketPlot}} that enforce axis labels
to be always perpendicular.
}
