% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{plot_colored_peptide_map}
\alias{plot_colored_peptide_map}
\title{Plot colored peptide map}
\usage{
plot_colored_peptide_map(
  prot,
  fill_by,
  min_y = 0,
  width_y = 0.025,
  step_y = 0.033,
  border_color = "white",
  border_size = NULL,
  aa_step = 20
)
}
\arguments{
\item{prot}{`data.frame` containing peptides from a single protein with
columns "First_AA" and "Last_AA".}

\item{fill_by}{character; name of a column in `prot` used to fill the peptide
rectangles.}

\item{min_y}{numeric; shifts peptides vertically.}

\item{width_y}{numeric; vertical width of each peptide. Must be less than
`step_y`.}

\item{step_y}{numeric; step size to resolve peptide overlaps.}

\item{border_color}{character; color used for the borders of the peptide
rectangles.}

\item{border_size}{numeric or `NULL`; linewidth of the borders around the
peptide rectangles. If `NULL` (default), the linewidth will be calculated
from the protein length.}

\item{aa_step}{numeric; distance between x axis breaks, starting at `aa_step`
and continuing along the length of the protein.}
}
\description{
Plot colored peptide map
}
