% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Dotplot of GO or Pathway Enrichment Results}
\usage{
plot_enrichment(
  x,
  num.categories = 10,
  sentence.case = TRUE,
  p.adjust.cutoff = 0.05,
  label.wrap.chars = Inf,
  point.size = 4,
  ...
)
}
\arguments{
\item{x}{an \code{\link[DOSE]{enrichResult-class}} object.}

\item{num.categories}{integer; the number of terms to plot. The default
selects the 10 terms with the highest counts, including ties.}

\item{sentence.case}{logical; whether to capitalize the first letter of each
term. Certain exceptions such as "mRNA" will not be capitalized.
Default is \code{TRUE}.}

\item{p.adjust.cutoff}{the significance cutoff for enrichment results. Only
terms with adjusted p-values less than \code{p.adjust.cuttoff} will be
considered for plotting.}

\item{label.wrap.chars}{the maximum number of characters per line for each
term.}

\item{point.size}{the size of the plotted points. Default is 4.}

\item{...}{additional arguments passed to
\code{\link[ggplot2]{scale_color_gradient}}.}
}
\description{
Create a dotplot of Gene Ontology or Pathway enrichment results
from \code{\link[ReactomePA]{enrichPathway}},
\code{\link[clusterProfiler]{enrichGO}},
or \code{\link[clusterProfiler]{enrichKEGG}}.
}
\examples{
library(clusterProfiler)

# Data from clusterProfiler
data(gcSample)

# GO: BP Enrichment results
x <- enrichGO(gene = gcSample[[6]],
              OrgDb = "org.Hs.eg.db",
              ont = "BP",
              pvalueCutoff = 1,
              universe = unique(unlist((gcSample))),
              qvalueCutoff = 1,
              minGSSize = 0)

# For the purposes of this example, the terms will not be
# filtered based on their adjusted p-values.
plot_enrichment(x, p.adjust.cutoff = 1,
                num.categories = 6, label.wrap.chars = 35,
                low = "darkred", high = "grey85")
}
