% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature.R
\name{plot_feature}
\alias{plot_feature}
\title{Plotting Individual Protein (Feature)}
\usage{
plot_feature(m, feature, feature_name_col = NULL, color_by = NULL,
  order_by = color_by)
}
\arguments{
\item{m}{MSnSet or ExpressionSet object}

\item{feature}{Name of the feature to plot. Should be in featureNames(m).}

\item{feature_name_col}{Name of the column in fData(m) that 
should contain name of the feature. 
The default value is NULL and matches feature
against featureNames(m).}

\item{color_by}{One of the varLabels(m) that will be used to color points.
Default is no coloring.}

\item{order_by}{One of the varLabels(m) for ordering points.
Defaul is `color_by`. If no ordering desired use NULL.}
}
\description{
The simplest way to plot an abundace profile.
}
\examples{
data(cptac_oca)
plot_feature(oca.set, "NP_001077422.1")
plot_feature(oca.set, "NP_001077422.1", color_by = "Batch")
plot_feature(oca.set, "NP_001077422.1", color_by = "Batch", order_by = "iTRAQ_ID")
plot_feature(oca.set, feature = "NP_001077422.1", feature_name_col = "RefSeq", color_by = "Batch")
}
