% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot}
\usage{
plot_pca(
  eset,
  phenotype = NULL,
  label = NULL,
  z_score = TRUE,
  show_ellipse = TRUE,
  components = 1:2,
  biplot = FALSE,
  biplot_labels = NULL,
  biplot_label_color = "blue",
  num_features = 5,
  show_NA = TRUE,
  legend_title = phenotype,
  ...
)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object.}

\item{phenotype}{\code{NULL} or string; one of \code{colnames(pData(eset))}.
This is used to color the points or labels, if \code{label} is not
\code{NULL}. Default is \code{NULL}, which colors all points or
labels black.}

\item{label}{\code{NULL} or string; one of \code{colnames(pData(eset))}.
If a string is provided, labels will be used instead of points.}

\item{z_score}{logical; whether to convert values to Z-Scores by sample.
Default is \code{TRUE}.}

\item{show_ellipse}{logical; whether to show the confidence ellipses if
\code{phenotype} is not \code{NULL}.}

\item{components}{numeric; a vector of length two specifying the principal
components to plot. Default is \code{c(1, 2)}, which plots PC1 on the
x-axis and PC2 on the y-axis. Order matters.}

\item{biplot}{logical; whether to display the biplot.}

\item{biplot_labels}{\code{NULL} or string; the name of a column in
\code{fData(eset)} used to label the biplot features. If \code{NULL}
(default), \code{featureNames(eset)} is used.}

\item{biplot_label_color}{string; the color of the feature labels. Default is
\code{"blue"}.}

\item{num_features}{numeric; the number of most influential features from
each principal component to label. Default is \code{5}.}

\item{show_NA}{logical; whether to include samples with missing
phenotype information. Default is \code{TRUE}.}

\item{legend_title}{string; title of the plot legend. Defaults to
\code{phenodata}.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_point}}
or \code{\link[ggplot2]{geom_text}}, such as \code{size} and \code{pch}.}
}
\value{
A ggplot object
}
\description{
A convenience function for plotting PCA scatter plot
for samples in ExpressionSet/MSnSet object. The biplot is essentially the
ggplot version of \code{\link[stats]{biplot.default}}
}
\examples{
library(MSnSet.utils)
data(cptac_oca)

# Default plot
plot_pca(oca.set)

# Color by SUBTYPE with custom legend title
plot_pca(oca.set, "SUBTYPE", legend_title = "Subtype")

# Color by SUBTYPE and label by Batch
plot_pca(oca.set, phenotype = "SUBTYPE", label = "Batch")

# Biplot
plot_pca(oca.set, biplot = TRUE)
}
