% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pecora.R
\name{plot_pecora}
\alias{plot_pecora}
\title{Plot PeCorA Results}
\usage{
plot_pecora(m, feature, treatment = NULL, median_mod = FALSE)
}
\arguments{
\item{m}{\code{MSnSet} object; output of
\code{\link[MSnSet.utils]{pecora_analysis}}.}

\item{feature}{character; the feature to plot. One of \code{featureNames(m)}.}

\item{treatment}{character; pattern used to search for columns of the form
"pecora_adj_pval_{treatment}" in \code{fData(m)}. If \code{NULL} (default),
this will automatically be determined if there is only one column that
starts with "pecora_adj_pval_".}

\item{median_mod}{logical; whether to use the median version of PeCorA. See
\code{?pecora_mod} for details.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot PeCorA Results
}
\details{
This function was inspired by the boxplots presented in the PeCorA
  publication (\url{https://doi.org/10.1021/acs.jproteome.0c00602}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
\seealso{
\code{\link[MSnSet.utils]{pecora_preprocess}},
  \code{\link[MSnSet.utils]{pecora_analysis}},
  \code{\link[MSnSet.utils]{pecora_mod}}
}
