% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{plot_peptide_on_protein_map}
\alias{plot_peptide_on_protein_map}
\title{Plot peptides on protein sequence}
\usage{
plot_peptide_on_protein_map(
  prot,
  fill_by,
  border_color = "white",
  border_size = NULL,
  aa_step = 20
)
}
\arguments{
\item{prot}{`data.frame` for exactly one protein with columns "ProtLen",
"ymax", "ymin", "First_AA", "Last_AA", and a column specified by `fill_by`.
`prot` should first be modified with `layout_manhattan` or
`layout_compact_stack`.}

\item{fill_by}{character; name of a column in `prot` used to fill the peptide
rectangles.}

\item{border_color}{character; color used for the borders of the peptide
rectangles.}

\item{border_size}{numeric or `NULL`; linewidth of the borders around the
peptide rectangles. If `NULL` (default), the linewidth will be calculated
from the protein length.}

\item{aa_step}{numeric; distance between x axis breaks, starting at `aa_step`
and continuing along the length of the protein.}
}
\description{
Plot peptides on protein sequence
}
