% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Create UMAP scatterplots of samples}
\usage{
plot_umap(
  msnset,
  phenotype = NULL,
  n_neighbors = NULL,
  min_dist = 0.1,
  n_epochs = 1000,
  point_size = 3,
  show_ellipse = TRUE,
  ellipse_type = "norm",
  ellipse_level = 0.5,
  show_na = TRUE,
  ...
)
}
\arguments{
\item{msnset}{an MSnSet object.}

\item{phenotype}{character; one of the \code{colnames(pData(msnset))}.}

\item{n_neighbors}{integer; number of nearest neighbors to use for UMAP.
If \code{NULL} (default) will use the closest integer to the
square root of the number of samples. If a vector of multiple
values are supplied, multiple plots will be generated and arranged
with \code{\link[ggplot2]{facet_wrap}}.}

\item{min_dist}{numeric; determines how close points appear in the final
layout (default: 0.1).}

\item{n_epochs}{integer; number of iterations performed during layout
optimization (default: 1000).}

\item{point_size}{numeric; size of the plotted points (default: 3).}

\item{show_ellipse}{logical; whether to plot ellipses.}

\item{ellipse_type}{character; the type of ellipse. "norm" (the default)
assumes a multivariate normal distribution, and "t" assumes a
multivariate t-distribution.}

\item{ellipse_level}{numeric; if \code{ellipse_type="euclid"}, the radius
of the circle to be drawn; otherwise, the confidence level
(default: 0.5).}

\item{show_na}{logical; whether points with missing phenotype labels
should be plotted (default: \code{TRUE}).}

\item{...}{Other arguments that can be passed to
\code{\link[umap]{umap.defaults}} or
\code{\link[ggplot2]{facet_wrap}}.}
}
\value{
A ggplot object
}
\description{
A convenience function for creating UMAP scatter plots of samples in an
ExpressionSet/MSnSet object.
}
\examples{
# Load msnset
data(srm_msnset)
# Do not color by phenotype
plot_umap(msnset)
# Plots for several values of n_neighbors
plot_umap(msnset, n_neighbors = c(6, 8, 10))
# Points colored by subject.type
plot_umap(msnset, phenotype = "subject.type",
          show_ellipse = TRUE)
# Plots for several values of n_neighbors with
# points colored by subject.type
plot_umap(msnset, phenotype = "subject.type",
          n_neighbors = c(6, 8, 10),
          min_dist = 0.01, show_ellipse = TRUE)
}
