% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(
  logFC,
  significance,
  sig_threshold = NULL,
  threshold_line_color = "red",
  features = NULL,
  top_n_features = 5,
  feature_labels = NULL,
  metadata = NULL,
  point_color = "grey",
  point_size = 3,
  use_theme = theme_bw(),
  ...
)
}
\arguments{
\item{logFC}{a numeric vector of the log2 fold-change for each feature.}

\item{significance}{a numeric vector of significance values
(p-value, q-value or adjusted p-value).}

\item{sig_threshold}{a numeric value specifying the significance threshold.
A dashed horizontal line will be drawn at this value. The default
is \code{NULL}.}

\item{threshold_line_color}{the color of the \code{sig_threshold} line.
Default is \code{"red"}.}

\item{features}{a character vector of feature names or \code{NULL} (default).}

\item{top_n_features}{the number of top significant features to label on
the plot if \code{features} is not \code{NULL}. Default is \code{5}.}

\item{feature_labels}{a character vector with the names of the features to
be labeled. Any features that should not be labeled must be \code{NA}.}

\item{metadata}{a data frame containing information that should be
accessible by additional layers. Useful for faceting.}

\item{point_color}{a color or a vector used to color the points.
Default is \code{"grey"}.}

\item{point_size}{the size of the points. Default is \code{3}.}

\item{use_theme}{the plot theme. Default is \code{theme_bw()}.}

\item{...}{additional arguments passed to
\code{\link[ggrepel]{geom_text_repel}}}
}
\value{
A ggplot object.
}
\description{
A convenience function for creating a volcano plot.
}
\examples{
library(MSnSet.utils)
data("cptac_oca")

res <- limma_a_b(oca.set,
                 model.str = "~ PLATINUM.STATUS + AGE",
                 coef.str = "PLATINUM.STATUS")

# Base plot
plot_volcano(logFC = res$logFC,
             significance = res$P.Value)

# Label top 10 most significant features
plot_volcano(logFC = res$logFC,
             significance = res$P.Value,
             sig_threshold = 0.05,
             features = rownames(res),
             top_n_features = 10) +
ylab("p-value") +
xlab(expression(log[2](paste(fold, "-", change))))

# Label top 5 most significant features and
# color by the sign of the average expression.
plot_volcano(logFC = res$logFC,
             significance = res$P.Value,
             sig_threshold = 0.05,
             features = rownames(res),
             point_color = factor(sign(res$AveExpr))) +
ylab("p-value") +
xlab(expression(log[2](paste(fold, "-", change))))
}
