% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_obsolete.R
\name{plot_volcano_obsolete}
\alias{plot_volcano_obsolete}
\title{Old Volcano Plot}
\usage{
plot_volcano_obsolete(
  logFC,
  significance,
  feature_names = NULL,
  threshold = NULL,
  top_n_names = 5,
  scale_xy = 1,
  ...
)
}
\arguments{
\item{logFC}{a numeric vector of log2 of fold change for each feature}

\item{significance}{a numeric vector of significance values
(p-value, q-value or adjusted p-value)}

\item{feature_names}{a character vector with the names of the features to
diplay or NULL (default)}

\item{threshold}{a numeric value where to draw a
significance threshold or NULL (default)}

\item{top_n_names}{number of top significant features
to show the names (default 5)}

\item{scale_xy}{numeric controls the spread on x vs y axis. Default 1.}

\item{...}{additional arguments passed to \code{\link[FField]{FFieldPtRep}}}
}
\value{
plot
}
\description{
This function is deprecated. Please use \code{\link{plot_volcano}} instead.
}
\examples{
# Not run
\dontrun{
library("MSnSet.utils")
data("cptac_oca")

res <- limma_a_b(oca.set,
                 model.str = "~ PLATINUM.STATUS + AGE",
                 coef.str = "PLATINUM.STATUS")

volcano_plot_obsolete(res$logFC,
                      res$P.Value,
                      feature_names = rownames(res),
                      threshold = 0.05,
                      top_n_names = sum(res$P.Value < 0.01, na.rm=TRUE),
                      rep.fact=10000, adj.lmt=1000, adj.max=1000) +
theme(text=element_text(size = 20))
}
}
