% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUC.R
\name{plotAUC}
\alias{plotAUC}
\title{Plot AUC}
\usage{
plotAUC(modelingResult)
}
\arguments{
\item{modelingResult}{output from either 
\code{lr_modeling} or \code{rf_modeling}}
}
\description{
Plot AUC after LOOCV Model Evaluation
}
