% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_NA.R
\name{postmean}
\alias{postmean}
\alias{it.sol}
\alias{postvar}
\title{These are helper functions for ComBat.NA
These functions perform the Bayesian adjustments to the batch effects
They do this in a iterative way.}
\usage{
postmean(g.hat, g.bar, n, d.star, t2)

it.sol(sdat, g.hat, d.hat, g.bar, t2, a, b, conv = 1e-04)

postvar(sum2, n, a, b)
}
\description{
Credit goes to the original authors of the sva package
}
\section{Functions}{
\itemize{
\item \code{it.sol}: 

\item \code{postvar}: 
}}

