% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_inference.R
\name{protein_inference}
\alias{protein_inference}
\title{Protein Inference}
\usage{
protein_inference(x, Accession = "Accession", PepSeq = "PepSeq")
}
\arguments{
\item{x}{table (data.frame, data.frame or tbl_df) with one column correspoding to peptides
and the other to proteins or in general case accessions}

\item{Accession}{character corresponding to column name with protein accesssions}

\item{PepSeq}{character corresponding to column name with peptide sequences}
}
\value{
data.frame with two additional columns
     \describe{
         \item{\code{UniqueAndRazor}}{TRUE if raw correspond to unique or razor peptide to protein mapping}
         \item{\code{JustifiedAccession}}{TRUE if protein has any unique peptide that justify its presence}
     }
}
\description{
A function for inferring parsimonious peptide to protein mapping.
}
\examples{
library(tibble)
x <- tribble(
    ~pep, ~prot,
    "AAAAK", "P1",
    "CCCCK", "P1",
    "LLLLK", "P1",
    "AAAAK", "P2",
    "DDDDK", "P2",
    "AAAAK", "P3")
protein_inference(x, Accession = "prot", PepSeq = "pep")

}
