% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaxQuantPeptideCounts.R
\name{readMaxQuantPeptideCounts}
\alias{readMaxQuantPeptideCounts}
\title{Reading MaxQuant Output at Peptide MS/MS Counts Level}
\usage{
readMaxQuantPeptideCounts(path = ".")
}
\arguments{
\item{path}{character path to the folder containing "peptides.txt file".
It should be in "{raw files folder}/combined/txt" of MaxQuant output.}
}
\value{
\code{MSnSet} object
}
\description{
Reads in "evidence.txt" output (or its compressed version) as MSnSet.
}
\note{
The "evidence.txt" file can be compressed (gzip) to save space.
}
\examples{

# to be added
}
