% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaxQuantPeptides.R
\name{readMaxQuantPeptides}
\alias{readMaxQuantPeptides}
\title{Reading MaxQuant Output}
\usage{
readMaxQuantPeptides(path = ".", msms_id_only = FALSE, verbose = 1)
}
\arguments{
\item{path}{character path to the folder containing "peptides.txt file".
It should be in "{raw files folder}/combined/txt" of MaxQuant output.}

\item{msms_id_only}{logical If TRUE retains values associated with MS/MS peptide
identifications only. Otherwise reports IDs identified both by MS/MS
and match between the runs (MBR). Default it FALSE - that is reporting
data from both type of identifications.}

\item{verbose}{numeric controls the text output}
}
\value{
\code{MSnSet} object
}
\description{
Reads in "peptides.txt" output (or its compressed version) as MSnSet.
}
\note{
The "peptides.txt" file can be compressed (gzip) to save space.

The quantitative data reported by this function is in "Intensity" columns of
      the "peptides.txt" file.
}
\examples{

# label-free data
m <- readMaxQuantPeptides(system.file("extdata/MaxQuantPep",
                                       package="MSnSet.utils"))
exprs(m) <- log2(exprs(m))
exprs(m) <- sweep(exprs(m), 1, rowMeans(exprs(m), na.rm=TRUE), '-')
image_msnset(m)

}
