% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readMaxQuantProtGroups.R
\name{readMaxQuantProtGroups}
\alias{readMaxQuantProtGroups}
\title{Reading MaxQuant Output}
\usage{
readMaxQuantProtGroups(path)
}
\arguments{
\item{path}{character path to the folder containing "proteinGroups.txt file".
It should be in "{raw files folder}/combined/txt"}
}
\value{
\code{MSnSet} object
}
\description{
Reads "proteinGroups.txt" output as MSnSet.
}
\details{
So far uses only LFQ data as quantitative data. As feature data, only first
12 columns and iBAQ columns are used. Features are based on the genes.
In case of multiple proteins matching the same gene, only the one with
the highest iBAQ value is returned.
}
\note{
The "proteinGroups.txt" file can be compressed to save space.
}
\examples{
m <- readMaxQuantProtGroups(system.file("extdata/MaxQuant",
                                         package="vp.misc"))
library("vp.misc")
exprs(m) <- log2(exprs(m))
exprs(m) <- sweep(exprs(m), 1, rowMeans(exprs(m), na.rm=TRUE), '-')
image_msnset(m)
}

