% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMaxQuantSummary.R
\name{readMaxQuantSummary}
\alias{readMaxQuantSummary}
\title{Reading MaxQuant Output}
\usage{
readMaxQuantSummary(path)
}
\arguments{
\item{path}{character path to the folder containing "summary.txt file".
It should be in "{raw files folder}/combined/txt"}
}
\value{
data.frame
}
\description{
Reads in "summary.txt" output.
}
\examples{

fpath <- list.files(path = path,
  pattern = "summary.txt",
  full.names = TRUE)
stopifnot(length(fpath) == 1)
smmr <- readMaxQuantSummary(fpath)
  
}
