% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SkyLine.R
\name{readSkyLinePRR}
\alias{readSkyLinePRR}
\title{Read SkyLine Results}
\usage{
readSkyLinePRR(file)
}
\arguments{
\item{file}{character, path to the file}
}
\value{
\code{\link[MSnbase]{MSnSet}} object
}
\description{
Reads SkyLine ouput file "Peptide Ratio Results.csv" as MSnSet
}
\examples{
\dontrun{#to be filled}

}

