% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_using_representative.R
\name{rollup_using_representative}
\alias{rollup_using_representative}
\title{Selecting one feature to represent a group of features}
\usage{
rollup_using_representative(msnset, roll_by, prioritize_by)
}
\arguments{
\item{msnset}{msnset (or most likely eset subclass) object}

\item{roll_by}{character. A column name in pData(msnset).
Controls grouping of the features.}

\item{prioritize_by}{character. Feature variables to prioritize the selection.
Note, ordering is always in the descending order.}
}
\value{
MSnSet object
}
\description{
The typical use of the function would be selecting representative peptide
per protein. Another use is selecting representative FAIMS CV value per
proteoform. This function is alternative to reference rollup (rrollup).
Instead of scaling peptides' intensities, here we select one peptide to
represent a protein.
}
\examples{
# Not run
\dontrun{
fData(m)$proteoform <- with(fData(m), paste(Gene, pcGroup, sep="_"))
fData(m)$count <- rowSums(!is.na(exprs(m)))
m2 <- rollup_using_representative(m)
}
}
