% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rrollup}
\alias{rrollup}
\title{Rolling-Up Features Abundance Measurements}
\usage{
rrollup(
  msnset,
  rollBy,
  rollFun,
  algorithm = c("reference", "sum"),
  verbose = TRUE
)
}
\arguments{
\item{msnset}{msnset (or most likely eset subclass) object}

\item{rollBy}{character. A column name in pData(msnset)}

\item{rollFun}{function for roll-up. "-" for log-transformed data,
"/" for normal scale}

\item{algorithm}{Either scaling to the most observed peptided ("reference")
or simply summing ("sum"). Summing makes sense only in
case of original label-free intensities.}

\item{verbose}{logical. Controls real-time output.}
}
\value{
MSnSet object
}
\description{
A primary purpose of this function is to roll-up peptide
measurements to protein level. Ultimately it can be used
for rolling-up or combining any features according to a grouping
factor.
}
\examples{
data(srm_msnset)
dim(msnset)
msnset2 <- rrollup(msnset, rollBy="Gene", rollFun="-", verbose=TRUE)
dim(msnset2)
}
