% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_args.R
\name{update_args}
\alias{update_args}
\title{Update a List of Arguments}
\usage{
update_args(old_list, new_list)
}
\arguments{
\item{old_list}{old list}

\item{new_list}{new list}
}
\value{
a list
}
\description{
Combine two lists. Used to update arguments in
\code{\link{plot_volcano}}, \code{\link{plot_pca}}, and
\code{\link{complex_heatmap}}. All list elements must be named.
}
\examples{

old_list <- list(color = "red", mapping = list(x = letters[1:3], y = 1:2))
new_list <- list(color = "blue", mapping = list(x = letters[4:6]))

update_args(old_list, new_list)
# returns list(color = "blue", mapping = list(x = letters[4:6], y = 1:2))

}
