% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable2.R
\name{xtable2}
\alias{xtable2}
\title{Customized xtable}
\usage{
xtable2(x, col.lab.rot = FALSE, ...)
}
\arguments{
\item{x}{typically a \code{data.frame}.}

\item{col.lab.rot}{describe this parameter.}

\item{...}{additional arguments passed to \code{link[xtable]{xtable}}.}
}
\value{
an object of class \code{"xtable"}
}
\description{
The main customization pertains to alternating
row shading. This one will work only in LaTeX-type documents.
See \code{\link[xtable]{xtable}} for details.
}
\seealso{
\code{\link[xtable]{xtable}}
}
