% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.plus.2.R
\name{heatmap.plus.2}
\alias{heatmap.plus.2}
\title{Tweak of "heatmap.plus"}
\usage{
heatmap.plus.2(
  x,
  Rowv = NULL,
  Colv = if (symm) "Rowv" else NULL,
  distfun = dist,
  hclustfun = hclust,
  reorderfun = function(d, w) reorder(d, w),
  add.expr,
  symm = FALSE,
  revC = identical(Colv, "Rowv"),
  scale = c("row", "column", "none"),
  na.rm = TRUE,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  propColSide = 0.1,
  propRowSide = 0.1,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  keep.dendro = FALSE,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted}

\item{Rowv}{determines if and how the *row* dendrogram should be computed and
reordered. Either a dendrogram or a vector of values used to reorder the
row dendrogram or \code{NA} to suppress any row dendrogram (and reordering)
or by default, \code{NULL}.}

\item{Colv}{determines if and how the *column* dendrogram should be
reordered. Has the same options as the Rowv argument above and additionally
when \code{x} is a square matrix, \code{Colv = "Rowv"} means that columns
should be treated identically to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity) between
both rows and columns. Defaults to \code{dist}.}

\item{hclustfun}{function used to compute the hierarchical clustering when
\code{Rowv} or \code{Colv} are not dendrograms. Defaults to \code{hclust}.}

\item{reorderfun}{function(d,w) of dendrogram and weights for reordering the
row and column dendrograms. The default uses \code{reorder.dendrogram}.}

\item{add.expr}{expression that will be evaluated after the call to
\code{image}. Can be used to add components to the plot.}

\item{symm}{logical indicating if \code{x} should be treated symmetrically;
can only be true when \code{x} is a square matrix.}

\item{revC}{logical indicating if the column order should be reversed for
plotting, such that e.g., for the symmetric case, the symmetry axis is as
usual.}

\item{scale}{character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. The default
is \code{"row"} if \code{symm} false, and \code{"none"} otherwise.}

\item{na.rm}{logical indicating whether \code{NA}'s should be removed.}

\item{margins}{numeric vector of length 2 containing the margins (see
\code{par(mar= *)}) for column and row names, respectively.}

\item{ColSideColors}{(optional) character matrix with number of columns
matching number of columns in \code{x}. Each column is plotted as a column
similar to heatmap()'s ColSideColors. colnames() will be used for labels if
present.}

\item{RowSideColors}{(optional) character matrix with number of rows matching
number of rows in \code{x}. Each column is plotted as a row similar to
heatmap()'s ColSideColors. colnames() will be used for labels if present.}

\item{propColSide}{numeric controls how much room to allocate for the side
color bars. E.g. 0.5 is even split between the heatmap and the bars.
Default is 0.1. That is, 10\% bars, 90\% heatmap.}

\item{propRowSide}{numeric see above}

\item{cexRow}{positive number, used as \code{cex.axis} in for the row axis
labeling. The default currently only uses number of rows.}

\item{cexCol}{see above.}

\item{labRow}{character vectors with row labels to use; defaults to
rownames().}

\item{labCol}{see above.}

\item{main}{main title; defaults to none.}

\item{xlab}{x-axis title; defaults to none.}

\item{ylab}{y-axis title; defaults to none.}

\item{keep.dendro}{logical indicating if the dendrogram(s) should be kept as
part of the result (when \code{Rowv} and/or \code{Colv} are not \code{NA}).}

\item{verbose}{logical indicating if information should be printed.}

\item{...}{additional arguments passed to \code{\link[graphics]{image}}.}
}
\description{
This function is basically
\href{https://github.com/cran/heatmap.plus/blob/master/R/heatmap.plus.R}{heatmap.plus::heatmap.plus}
from \href{https://github.com/cran/heatmap.plus/blob/master}{heatmap.plus}
package, but it includes a pair of additional arguments controlling the
proportion of row and column color sidebars.
}
\examples{
z = matrix(rnorm(30),nrow=5,ncol=6);
rlab = matrix(as.character(c(1:5,2:6,3:7,4:8)),nrow=5,ncol=4);
clab = matrix(as.character(c(1:6,6:1)),nrow=6,ncol=2);
colnames(rlab) = LETTERS[1:dim(rlab)[2]];
colnames(clab) = 1:dim(clab)[2];
heatmap.plus.2(z, ColSideColors=clab,
                  RowSideColors=rlab,
                  propColSide=0.1,
                  propRowSide=0.2);
}
