% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_a_b.R
\name{limma_a_b}
\alias{limma_a_b}
\alias{limma_gen}
\alias{limma_contrasts}
\title{Wrapper for Limma}
\usage{
limma_a_b(eset, model.str, coef.str, ...)

limma_gen(eset, model.str, coef.str, ...)

limma_contrasts(
  eset,
  model.str,
  coef.str,
  contrasts,
  adjust.method = "BH",
  remove_name = TRUE,
  ...
)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{model.str}{character formulation of the model (e.g. "~ a + b")}

\item{coef.str}{character coefficient of interest. E.g. "a".
Either numeric or factor with 2 levels.}

\item{...}{arguments for \code{\link[limma]{lmFit}}}

\item{ref.str}{(character) Reference string for pairwise comparisons}
}
\value{
data.frame. Basically output of 
     \code{\link[limma]{topTable}} function.
}
\description{
A convenience wrapper for limma
}
\section{Functions}{
\itemize{
\item \code{limma_gen}: More generalized approach. The tested variable doesn't have to be a
factor with 2 levels like in limma_a_b.

\item \code{limma_contrasts}: For multiple pairwise comparisons. The tested variable doesn't have to be a
factor with 2 levels like in limma_a_b.
}}

\note{
limma_a_b is for either 2-factor ANOVA or linear regression.
      limma_gen is a more generic version and will return results
      for all the coefficients that match \code{coef.str} pattern.
}
\examples{
library("MSnSet.utils")
data("cptac_oca")
library("limma")
ee <- oca.set
ee <- ee[rowSums(!is.na(exprs(ee))) >= 30,]
model.string <- "~ PLATINUM.STATUS + AGE" # model with covariates
coef.string <- "PLATINUM.STATUS" # note, only two levels allowed here

res <- limma_a_b(oca.set, 
                 model.str = "~ PLATINUM.STATUS + AGE", 
                 coef.str = "PLATINUM.STATUS")
head(res)
}
