% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oca.set}
\alias{oca.set}
\title{CPTAC Ovarian Cancer Proteomics Dataset}
\format{
An \code{MSnSet} with 8103 features and 73 samples.
}
\source{
CPTAC study
}
\usage{
data("cptac_oca")
}
\description{
Processed iTRAQ data from both PNNL and JHU sites.
}
\section{pData}{

A data frame consisting of 73 rows (samples) and 39 columns:
\describe{
  \item{Tumor.Nuclei}{Percent of cells in the tissue sample with tumor
  nuclei.}
  \item{Necrosis}{Percent of tumor nuclei that are necrotic.}
  \item{PNNL.Weight..mg.}{Weight of the non-tumor tissue in milligrams.}
  \item{Residual.Tumor.Tissue..mg.}{Weight of the tumor tissue in
  milligrams.}
  \item{TSS}{}
  \item{primary_therapy_outcome_success}{One of "COMPLETE.RESPONSE",
  "STABLE.DISEASE", or "PROGRESSIVE.DISEASE".}
  \item{days_to_death}{Days between }
  \item{Platinum.Refractory}{}
  \item{tumor_grade}{}
  \item{tumor_residual_disease}{}
  \item{tumor_stage}{Tumor stage.}
  \item{anatomic_organ_subdivision}{Either "left", "right", or "bilateral".}
  \item{days_to_birth}{Negative age (in days) of subject.}
  \item{days_to_tumor_recurrence}{Days before first tumor recurrence.}
  \item{race}{The race of the subject.}
  \item{site_of_tumor_first_recurrence}{"LOCO.REGIONAL" or "METASTASIS".}
  \item{Batch}{The sample batch.}
  \item{additional_chemo_therapy}{"YES" if the subject received additional
  chemotherapy; "NO" otherwise.}
  \item{additional_drug_therapy}{"YES" if the subject received additional
  drug therapy; "NO" otherwise.}
  \item{additional_pharmaceutical_therapy}{"YES" if the subject received
  additional pharmaceutical therapy; "NO" otherwise.}
  \item{additional_radiation_therapy}{"YES" if the subject received
  additional radiation therapy; "NO" otherwise.}
  \item{targeted_molecular_therapy}{"YES" if the subject received
  targeted molecular therapy; "NO" otherwise.}
  \item{year_of_initial_pathologic_diagnosis}{The year when subjects were
  first diagnosed with ovarian cancer.}
  \item{iTRAQ_Batch}{The iTRAQ batch identifier.}
  \item{iTRAQ_ID}{}
  \item{ReporterIon}{}
  \item{SUBTYPE}{Ovarian cancer subtype. Either "Differentiated",
  "Immunoreactive", "Mesenchymal", or "Proliferative".}
  \item{AGE}{Subject age in years.}
  \item{PLATINUM.STATUS}{The tumor's response to platinum therapy. Either
  "SENSITIVE" or "RESISTANT".}
  \item{TUMORRESIDUALDISEASE}{}
  \item{SURVIVALSTATUS}{The survival status of the patient. Either "LIVING"
  or "DECEASED".}
  \item{SURVIVALMONTHS}{}
  \item{RECURRENCE.MONTHS}{}
  \item{Differentiated.binary}{\code{TRUE} if \code{SUBTYPE} is
  "Differentiated"; \code{FALSE} otherwise.}
  \item{Immunoreactive.binary}{\code{TRUE} if \code{SUBTYPE} is
  "Immunoreactive"; \code{FALSE} otherwise.}
  \item{Mesenchymal.binary}{\code{TRUE} if \code{SUBTYPE} is
  "Mesenchymal"; \code{FALSE} otherwise.}
  \item{Proliferative.binary}{\code{TRUE} if \code{SUBTYPE} is
  "Proliferative"; \code{FALSE} otherwise.}
  \item{SILHOUETTE.WIDTH}{}
  \item{fake}{}
}
}

\section{fData}{

\describe{
  \item{RefSeq}{The NCBI reference sequence.}
}
}

\examples{
data("cptac_oca")
head(pData(oca.set))
head(fData(oca.set))
head(exprs(oca.set))
}
\keyword{datasets}
