% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_analysis.R
\name{pecora_analysis}
\alias{pecora_analysis}
\title{Performs the PeCorA analysis described in https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/.
Take a look at the linked paper to get an idea of the procedure.}
\usage{
pecora_analysis(m, treatment_string, proteins, median_mod = FALSE)
}
\arguments{
\item{m}{msnset object containing a pData AND fData table. fData must have a Protein and Peptide column.}

\item{treatment_string}{the name of the column in pData containing treatment information. The data can be numerical of categorical.}

\item{proteins}{character vector indicating which proteins to analyze.}

\item{median_mod}{logical for whether to use the original PeCorA (median_mod = FALSE), or the median version (median_mod = TRUE). See the help of PeCorA_mod for more details.}
}
\value{
PeCorA results table for the supplied proteins. Saves plots of the significant peptides to the given folder.
}
\description{
Takes an msnset, a list of proteins, and a treatment variable. msnset is assumed to be normalized, for instance it can be the output of
pecora_preprocessing, which normalizes as in the original PeCorA paper.
The PeCorA function is then used for the analysis. Returns a table containing Peptide, Protein, and significance.
}
