% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeCorA_mod.R
\name{PeCorA_mod}
\alias{PeCorA_mod}
\title{This is a modified version of the PeCorA function.
For the original paper see here: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/.}
\usage{
PeCorA_mod(t, median_mod = FALSE)
}
\arguments{
\item{t}{PeCorA input table. See the original PeCorA function for details.}

\item{median_mod}{logical for whether to use the median version of PeCorA (median_mod = TRUE), or the original (median_mod = FALSE)}
}
\value{
PeCorA results table.
}
\description{
The modification involves aggregating the intensities of "all other peptides" within
the test described in the PeCorA paper. We aggregate "all other peptides" by sample using the median, this
way each peptide mapping to a particular protein is not treated as an 'independent' observation.
The effect is that p-values are much higher (less significant) than in the original version of PeCorA.
}
\details{
This median modification is turned OFF by default, and only affects the p-values reported by PeCorA.
}
