% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_preprocess.R
\name{pecora_preprocess}
\alias{pecora_preprocess}
\title{Preprocess Protein Expression Data for PeCorA}
\usage{
pecora_preprocess(
  m,
  treatment,
  control_group = NULL,
  standardize_samples = TRUE,
  center_peptides = TRUE
)
}
\arguments{
\item{m}{\code{MSnSet} object with pData and fData tables. fData must have
columns "Protein" and "Peptide".}

\item{treatment}{character; the name of the pData column containing treatment
information. The data can be numeric or categorical.}

\item{control_group}{character; the control group, if \code{treatment} is
categorical, this is one of the levels.}

\item{standardize_samples}{logical; whether to standardize samples (columns)
in exprs data.}

\item{center_peptides}{logical; whether to mean-center each peptide. If the
data in the \code{treatment} column is of type "character" or "factor",
peptides will be zero-centered on the control group mean. Otherwise, the
mean of all samples are used for centering.}
}
\value{
An \code{MSnSet} object with modified exprs data.
}
\description{
This function optionally standardizes samples and/or
  zero-centers peptides using the overall mean of each peptide or the mean of
  the \code{control_group} samples.
}
\details{
If both \code{standardize_samples} and \code{center_peptides} are
  set to \code{FALSE}, the MSnSet will not be modified, and this step can be
  safely skipped.

  This function was based on the \code{PeCorA_preprocessing} function from
  the PeCorA R package (\url{https://github.com/jessegmeyerlab/PeCorA}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
\seealso{
\code{\link[MSnSet.utils]{normalizeByGlob}},
  \code{\link[MSnSet.utils]{pecora_analysis}},
  \code{\link[MSnSet.utils]{pecora_mod}},
  \code{\link[MSnSet.utils]{plot_pecora}}
}
