% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_significance_map.R
\name{peptide_significance_map}
\alias{peptide_significance_map}
\title{Peptide significance map}
\usage{
peptide_significance_map(
  x,
  accession,
  p_val_from,
  type = c("discrete", "continuous", "y"),
  name_from = c("accession"),
  ...
)
}
\arguments{
\item{x}{\code{MSnSet} object.}

\item{accession}{\code{character}. Name of the protein
(or accession in general terms) to plot peptides for.}

\item{p_val_from}{character. Name of the column in `fData` with p-values.}

\item{type}{character. Different ways of representing p-values.}

\item{name_from}{character. Column in `fData` that `accession` is taken from.
Typically it is "accession".}

\item{...}{to be documented}
}
\value{
\code{ggplot} object
}
\description{
Mapping peptides onto the protein sequence with showing statisical significance.
}
\examples{

# need to add peptide-level MSnSet

}
