% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca_v1}
\alias{plot_pca_v1}
\alias{plot_pca_v2}
\alias{plot_pca_v3}
\alias{plot_pca_v4}
\title{PCA Plot}
\usage{
plot_pca_v1(
  eset,
  phenotype = NULL,
  show.ellipse = TRUE,
  show.NA = TRUE,
  legend.title.width = 20
)

plot_pca_v2(eset, phenotype = NULL, names = FALSE)

plot_pca_v3(
  eset,
  phenotype = NULL,
  label = NULL,
  point_size = 5,
  show.ellipse = TRUE,
  show.NA = TRUE,
  legend.title.width = 20,
  ...
)

plot_pca_v4(
  eset,
  phenotype = NULL,
  label = NULL,
  point_size = 5,
  show.ellipse = TRUE,
  show.NA = TRUE,
  legend.title.width = 20,
  pc.index = c(1, 2),
  ...
)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{phenotype}{character one of the \code{colnames(pData(eset))}}

\item{show.ellipse}{logical determining to plot 95\% CI based on
Hotelling's T-test or not.}

\item{show.NA}{logical whether the datapoints for which phenotype is
unknown be shown. Default is \code{TRUE}.}

\item{legend.title.width}{integer Wrapping up too long legend titles.
Passed to stringr::str_wrap as width argument. Default is 20.}

\item{names}{this parameter is specified, but not actually used.}

\item{label}{character vector of point labels.}

\item{point_size}{numeric; the size of the plotted points.}

\item{...}{additional arguments passed to
\code{\link[ggrepel]{geom_label_repel}}.}

\item{pc.index}{numeric vector of principal components to plot. Defaults to
\code{c(1,2)}, so PC1 and PC2 will be used as the axes.}
}
\value{
A ggplot object
}
\description{
A convenience function for plotting PCA scatter plot
for samples in ExpressionSet/MSnSet object.
}
\section{Functions}{
\itemize{
\item \code{plot_pca_v2}: Alternative PCA

\item \code{plot_pca_v3}: Alternative PCA

\item \code{plot_pca_v4}: Alternative PCA with option to select pair of principal components
}}

\examples{
data(srm_msnset)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellipse = FALSE)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellipse = TRUE)
plot_pca_v1(msnset)
data(srm_msnset)
plot_pca_v2(msnset, phenotype = "subject.type")
plot_pca_v2(msnset)
data(srm_msnset)
plot_pca_v3(msnset, phenotype = "subject.type")
plot_pca_v3(msnset, phenotype = "subject.type", label = "sample.id")
plot_pca_v3(msnset)
data(srm_msnset)
plot_pca_v4(msnset, phenotype = "subject.type", pc.index=c(1,3))
plot_pca_v4(msnset, phenotype = "subject.type", label = "sample.id", pc.index=c(1,3))
plot_pca_v4(msnset, pc.index=c(1,3))
}
