% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca_v1}
\alias{plot_pca_v1}
\title{PCA Plot}
\usage{
plot_pca_v1(eset, phenotype = NULL, show.ellispe = TRUE, names = FALSE)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{phenotype}{character one of the \code{colnames(pData(eset))}}

\item{show.ellipse}{logical determining to plot 95\% CI based on
Hotelling's T-test or not.}
}
\value{
plot
}
\description{
A convenience function for plotting PCA scatter plot
for samples in ExpressionSet/MSnSet object.
}
\examples{
data(srm_msnset)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellispe = F)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellispe = T)
plot_pca_v1(msnset)
}

