% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca_v1}
\alias{plot_pca_v1}
\alias{plot_pca_v2}
\alias{plot_pca_v3}
\title{PCA Plot}
\usage{
plot_pca_v1(eset, phenotype = NULL, show.ellispe = TRUE, show.NA = TRUE,
  legend.title.width = 20)

plot_pca_v2(eset, phenotype = NULL, names = FALSE)

plot_pca_v3(eset, phenotype = NULL, label = NULL, show.ellispe = TRUE,
  show.NA = TRUE, legend.title.width = 20)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{phenotype}{character one of the \code{colnames(pData(eset))}}

\item{show.NA}{logical Should the datapoints for which phenotype is 
unknown be shown? Default is TRUE}

\item{legend.title.width}{integer Wrapping up too long legend titles.
Passed to stringr::str_wrap as width argument. Default is 20.}

\item{show.ellipse}{logical determining to plot 95\% CI based on 
Hotelling's T-test or not.}
}
\value{
plot
}
\description{
A convenience function for plotting PCA scatter plot
for samples in ExpressionSet/MSnSet object.
}
\section{Functions}{
\itemize{
\item \code{plot_pca_v2}: Alternative PCA

\item \code{plot_pca_v3}: Alternative PCA
}}

\examples{
data(srm_msnset)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellispe = F)
plot_pca_v1(msnset, phenotype = "subject.type", show.ellispe = T)
plot_pca_v1(msnset)
data(srm_msnset)
plot_pca_v2(msnset, phenotype = "subject.type")
plot_pca_v2(msnset)
data(srm_msnset)
plot_pca_v3(msnset, phenotype = "subject.type")
plot_pca_v3(msnset, phenotype = "subject.type", label = "subject.ID")
plot_pca_v3(msnset)
}
