% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PeCorA.R
\name{plot_PeCorA}
\alias{plot_PeCorA}
\title{Plot PeCorA Results}
\usage{
plot_PeCorA(
  m,
  peptide,
  protein,
  treatment = NULL,
  protein_column = "Protein",
  peptide_column = "Peptide",
  median_mod = TRUE
)
}
\arguments{
\item{m}{\code{MSnSet} object; output of \code{\link[MSnSet.utils]{PeCorA}}.}

\item{peptide}{character; the peptide of interest to plot. One of
\code{fData(m)[[peptide_column]]}.}

\item{protein}{character; the protein that contains \code{peptide} as one of
its peptides. One of \code{fData(m)[[protein_column]]}.}

\item{treatment}{character; pattern used to search for columns of the form
"pecora_adj_pval_{treatment}" in \code{fData(m)}. If \code{NULL} (default),
this will be automatically determined if there is only one column that
starts with "pecora_adj_pval_".}

\item{protein_column}{character; the name of a column in \code{fData(m)}
containing protein information. Default is "Protein".}

\item{peptide_column}{character; the name of a column in \code{fData(m)}
containing peptide information. Default is "Peptide".}

\item{median_mod}{logical; whether to use a modified version of PeCorA. See
details.}
}
\value{
A \code{ggplot} object.
}
\description{
Create boxplots (categorical \code{treatment}) or a scatterplot
  (numeric \code{treatment}) of PeCorA results for a given peptide.
}
\details{
This function was inspired by the boxplots presented in the PeCorA
  publication (\url{https://doi.org/10.1021/acs.jproteome.0c00602}).
}
\references{
Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G.
  (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential
  Proteoform Regulation. \emph{Journal of proteome research, 20(4)},
  1972–1980. \url{https://doi.org/10.1021/acs.jproteome.0c00602}
}
\seealso{
\code{\link[MSnSet.utils]{PeCorA}}
}
