% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP Plot}
\usage{
plot_umap(
  eset,
  phenotype = NULL,
  standardize = FALSE,
  pca = TRUE,
  n_perm = 10,
  n_neighbors = round(sqrt(length(sampleNames(eset)))),
  n_epochs = 1000,
  min_dist = 0.01,
  legend_title = NULL,
  legend_title_width = 20,
  point_size = 4,
  show_ellipse = TRUE,
  ellipse_type = "norm",
  show_na = TRUE,
  ...
)
}
\arguments{
\item{eset}{eset (or most likely eset subclass) object}

\item{phenotype}{character; one of the \code{colnames(pData(eset))}}

\item{standardize}{logical; should the rows of the expression data be
standardized? (default: \code{FALSE})}

\item{pca}{logical; whether PCA should be performed prior to UMAP
(default: \code{TRUE}).}

\item{n_perm}{integer; if pca is \code{TRUE}, this is the number of times
that the expression matrix should be permuted (default: 10).
This is done to determine the optimal number of principal
components to use.}

\item{n_neighbors}{integer; number of nearest neighbors to use for UMAP
(default: \eqn{\sqrt(n samples)}).}

\item{n_epochs}{integer; number of iterations performed during layout
optimization.}

\item{min_dist}{numeric; determines how close points appear in the final
layout (default: 0.01).}

\item{legend_title}{character; the title for the legend. If NULL, is set to
the wrapped up phenotype.}

\item{legend_title_width}{integer; wrapping up too long legend titles.
Passed to \code{stringr::str_wrap()} as width argument
(default: 20).}

\item{point_size}{numeric; size of the plotted points (default: 4).}

\item{show_ellipse}{logical; whether to plot 50\% confidence ellipses.}

\item{ellipse_type}{character; the type of ellipse. The default "norm"
assumes a multivariate normal distribution, and "t" assumes a
multivariate t-distribution.}

\item{show_na}{logical; should the data points for which phenotype is
unknown be shown? (default: \code{TRUE})}

\item{...}{Other arguments that can be passed to the UMAP configuration.
See \code{\link[umap]{umap.defaults}} for more details.}
}
\value{
A ggplot object
}
\description{
A convenience function for creating a UMAP scatter plot
for samples in an ExpressionSet/MSnSet object.
}
\examples{
data(srm_msnset)
plot_umap(msnset,
          phenotype = "subject.type",
          show_ellipse = F)
plot_umap(msnset,
          phenotype = "subject.type",
          min_dist = 0.5,
          show_ellipse = T)
plot_umap(msnset,
          phenotype = "subject.type",
          pca = F,
          show_ellipse = T)
plot_umap(msnset)

}
