% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(
  df,
  logFC,
  pvals,
  sig_threshold = NULL,
  label = NULL,
  num_features = 8L,
  point_args = list(),
  label_args = list()
)
}
\arguments{
\item{df}{\code{data.frame} or object that can be coerced to a
\code{data.frame}}

\item{logFC}{character; the name of the column in \code{df} containing log2
fold-changes.}

\item{pvals}{character; the name of the column in \code{df} containing
p-values (adjusted or not).}

\item{sig_threshold}{\code{NULL} (default) or numeric between 0 and 1; the
threshold indicating statistical significance. Creates a dashed horizontal
line at that value. Any points above this line in the graph are
statistically significant.}

\item{label}{\code{NULL} or character; the name of the column in \code{df}
used to label the top most significant features. See details for more.}

\item{num_features}{numeric; the number of most significant features to
label. Default is 8.}

\item{point_args}{a list of arguments passed to
\code{\link[ggplot2]{geom_point}}.}

\item{label_args}{a list of arguments passed to
\code{\link[ggrepel]{geom_label_repel}}.}
}
\value{
A ggplot object.
}
\description{
A convenience function for creating a volcano plot.
}
\details{
\code{sig_threshold} will create a secondary axis if the threshold
  is not one of the existing y-axis breaks.

  To label specific features (not top \code{num_features}), \code{label}
  needs to be the name of a column where all but the features that will be
  labeled are \code{NA}. Also, set \code{num_features} to \code{nrow(df)}.
}
\examples{
library(ggplot2) # additional plot modifications
library(MSnSet.utils)
data("cptac_oca")

# Differential analysis
df <- limma_a_b(oca.set,
                model.str = "~ PLATINUM.STATUS + AGE",
                coef.str = "PLATINUM.STATUS")
df$features <- rownames(df) # Add feature names column

# Base plot
p <- plot_volcano(df, logFC = "logFC", pvals = "P.Value")
p

# Change y-axis title
p + labs(y = "Unadjusted p-value")

# Add dashed line at y = 0.05
plot_volcano(df, logFC = "logFC", pvals = "P.Value",
             sig_threshold = 0.05)

# Label top 8 most significant features
plot_volcano(df, logFC = "logFC", pvals = "P.Value",
             label = "features")

# Change point opacity, point size, and color of labels
plot_volcano(df, logFC = "logFC", pvals = "P.Value",
             label = "features",
             point_args = list(alpha = 0.3, size = 3),
             label_args = list(color = "royalblue"))

}
