% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_NA.R
\name{postmean}
\alias{postmean}
\alias{it.sol}
\alias{postvar}
\title{Helper functions for ComBat.NA}
\usage{
postmean(g.hat, g.bar, n, d.star, t2)

it.sol(sdat, g.hat, d.hat, g.bar, t2, a, b, conv = 1e-04)

postvar(sum2, n, a, b)
}
\arguments{
\item{g.hat}{additive batch effect}

\item{g.bar}{additive batch effect mean}

\item{n}{number of samples in batch}

\item{d.star}{bayes adjusted multiplicative effect}

\item{t2}{additive batch effect variance}

\item{sdat}{standardized data}

\item{d.hat}{multiplicative batch effect}

\item{a}{shape parameter}

\item{b}{rate parameter}

\item{conv}{relative change threshold, determines when to stop iteration}

\item{sum2}{internal variable, used in Bayesian estimation of batch effects}
}
\description{
These functions perform the Bayesian adjustments to the batch effects
They do this in a iterative way.

Credit goes to the original authors of the sva package.
}
\section{Functions}{
\itemize{
\item \code{it.sol}: 

\item \code{postvar}: 
}}

