% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xtable2.R
\name{xtable2}
\alias{xtable2}
\title{Customized xtable}
\usage{
xtable2(x, col.lab.rot = FALSE, ...)
}
\arguments{
\item{x}{typically a \code{data.frame}.}
}
\value{
an object of class \code{"xtable"}
}
\description{
The main customization pertains to alternating
row shading. See \code{\link[xtable]{xtable}} for details.
}
\seealso{
\code{\link[xtable]{xtable}}
}

